:zap: v0.9.0 see ./changelog.md
Squashed commit of the following:

commit 6231f14d40abf5b19f345da88f6b74272f635c2a
Author: Jan Andrle <andrle.jan@centrum.cz>
Date:   Tue Jan 6 11:12:18 2026 +0100

    :abc:

commit 0d3e4b82ddd8081886fdca3b7b96e16c1885aa33
Author: Jan Andrle <andrle.jan@centrum.cz>
Date:   Tue Jan 6 10:59:08 2026 +0100

    :bug: packge rename → info.name

commit faf0bbdb817760d113413cdb0c1458701b365e64
Author: Jan Andrle <andrle.jan@centrum.cz>
Date:   Tue Jan 6 10:56:22 2026 +0100

    :bug:

commit 486b82b541d3311ff2a8cbc091703c41d8994e88
Author: Jan Andrle <andrle.jan@centrum.cz>
Date:   Tue Jan 6 10:44:03 2026 +0100

    :zap: :abc: Provides helpes via “extensions” and docs

    Squashed commit of the following:

    commit 1ff7d7b580e30872ccbe4055d87ca8c6cca1ee51
    Author: Jan Andrle <andrle.jan@centrum.cz>
    Date:   Tue Jan 6 10:43:27 2026 +0100

        :abc:

    commit 157b0308e6081d8a4b97bd572cd53466a6c9d1cb
    Author: Jan Andrle <andrle.jan@centrum.cz>
    Date:   Tue Jan 6 10:41:10 2026 +0100

        :zap:

    commit 15c3e677560e146e533ebe671dcd897e8390afc1
    Author: Jan Andrle <andrle.jan@centrum.cz>
    Date:   Tue Jan 6 10:13:19 2026 +0100

        :zap: :abc:

    commit 3712ed07f4123e3a7fabc538602f6c4f9e712b59
    Author: Jan Andrle <andrle.jan@centrum.cz>
    Date:   Mon Jan 5 16:48:21 2026 +0100

        :bug: :abc:

    commit 2eb4018f2f654b946cefbe1f59444f1736d52630
    Author: Jan Andrle <andrle.jan@centrum.cz>
    Date:   Mon Jan 5 16:30:10 2026 +0100

        :zap: improvements

    commit 58081c85365cf2ea90e47a1f93bfc05f2b4c5159
    Author: Jan Andrle <andrle.jan@centrum.cz>
    Date:   Fri Dec 5 14:32:31 2025 +0100

        :zap: Extensions instead of build-in

    commit d6330f8c1c39d14275f8c398de51218ad5c60d7b
    Author: Jan Andrle <andrle.jan@centrum.cz>
    Date:   Fri Nov 21 14:51:13 2025 +0100

        :zap: (non-public) allow completion[.extension]

    commit 1b523a42d1718f38f62e9c575e64d2fad0d90638
    Author: Jan Andrle <andrle.jan@centrum.cz>
    Date:   Fri Nov 21 13:00:13 2025 +0100

        :zap: `.templates .wget`

    commit 484fc86d744265a2e166f10011df8fe54229b882
    Author: Jan Andrle <andrle.jan@centrum.cz>
    Date:   Fri Nov 21 11:38:44 2025 +0100

        :zap: `.templates .mkdir`

    commit 45a1ce7715381d38e6a74e9c72af851cb55ffd55
    Author: Jan Andrle <andrle.jan@centrum.cz>
    Date:   Fri Nov 21 11:32:18 2025 +0100

        :zap: cleanup/minimize `.templates`

    commit cb538c50c51b97109b9554a347d72d0b5ece3c9f
    Author: Jan Andrle <andrle.jan@centrum.cz>
    Date:   Thu Nov 20 15:40:36 2025 +0100

        :abc:

    commit 8fc0ec36110e63ab852cf87a236abf536d0b0095
    Author: Jan Andrle <andrle.jan@centrum.cz>
    Date:   Thu Nov 20 15:33:36 2025 +0100

        :abc: :zap:

    commit f6506182f75959e014b9c68415b21192c4b34598
    Author: Jan Andrle <andrle.jan@centrum.cz>
    Date:   Thu Nov 20 15:23:06 2025 +0100

        :zap: subfolders and completion

    commit 24d9caf1a2346f0d3219730de23681e55541b629
    Author: Jan Andrle <andrle.jan@centrum.cz>
    Date:   Thu Nov 20 14:57:23 2025 +0100

        :zap: `.templates` improvements

    commit f8e68af9dd3e9c6e097ff43d7a51981c3e684e68
    Author: Jan Andrle <andrle.jan@centrum.cz>
    Date:   Wed Nov 19 15:02:43 2025 +0100

        :tada:

commit a5f351e25a1d8b90a7cb291f1658dc29d95c1dea
Author: Jan Andrle <andrle.jan@centrum.cz>
Date:   Fri Nov 14 17:04:37 2025 +0100

    :abc:

commit 21ed8e31343af2339a86de6579b3aed80ed1d116
Author: Jan Andrle <andrle.jan@centrum.cz>
Date:   Fri Nov 14 16:29:03 2025 +0100

    :zap: `.grep`

commit 39d63b1381306c8820c0b7c2fad40004a906acaa
Author: Jan Andrle <andrle.jan@centrum.cz>
Date:   Fri Nov 14 16:27:45 2025 +0100

    :abc:

commit 08dceedaf7685a349df6027108ceab263ced246b
Author: Jan Andrle <andrle.jan@centrum.cz>
Date:   Fri Nov 14 16:25:24 2025 +0100

    :abc:

commit cd6eba6d1a21a57a6afc5856c3e2fdd3a0b002f5
Author: Jan Andrle <andrle.jan@centrum.cz>
Date:   Fri Nov 14 16:23:08 2025 +0100

    :zap: :abc:

commit 89594bff5fb7be36490a933467a214c8d173d392
Author: Jan Andrle <andrle.jan@centrum.cz>
Date:   Fri Nov 14 13:39:32 2025 +0100

    :tv: :bug: build

commit 43ad816dec453f55a7f69383ce5ad3efe5b556b2
Author: Jan Andrle <andrle.jan@centrum.cz>
Date:   Fri Nov 14 12:59:55 2025 +0100

    :zap: ui

commit a089800cc96c0a490bd8d42a161318c02482b4d5
Author: Jan Andrle <andrle.jan@centrum.cz>
Date:   Thu Nov 13 15:53:17 2025 +0100

    :zap: cli refact and improvements

    - resolves passed files
    - consolidates log head

commit e329f8bbd714c8cea76e8a9c5d4bd341c3a3541f
Author: Jan Andrle <andrle.jan@centrum.cz>
Date:   Tue Aug 19 16:11:02 2025 +0200

    :zap: extracts find-bs

commit 23553742493e8672a4ee8c25362b56a0bf63f444
Author: Jan Andrle <andrle.jan@centrum.cz>
Date:   Tue Aug 19 15:45:17 2025 +0200

    :abc:

commit 6c757d06b3c0a9a8a1c27f371ca34a6b4fa9c5d2
Author: Jan Andrle <andrle.jan@centrum.cz>
Date:   Tue Jul 29 16:13:27 2025 +0200

    :zap: tada

+++ bs.js
@@ -3,130 +3,100 @@
 const // global functions/utils
 	{ log, format }= require("css-in-console"),
 	{ readFileSync, existsSync, mkdirSync, writeFileSync, statSync, fstatSync }= require("node:fs"),
-	{ join }= require("node:path"),
-	{ stdout, argv }= require("node:process"),
+	{ homedir } = require("node:os"),
+	{ join, relative }= require("node:path"),
+	{ stdout, argv, exit, cwd: pwd, chdir, kill, pid }= require("node:process"),
 	{ pipe,
-	  linesToMaxLength, passBuildArgs,
+	  passBuildArgs,
 	  listExecutables, isExecutable }= require("./src/utils.js"),
-	{ cwd: pwd }= require("node:process");
+	{ folderRoot, folderConfig }= require("./src/find-bs.js");
 const stdoutIsFIFO= (()=> {
 	try { return fstatSync(stdout.fd).isFIFO(); }
 	catch(_){ return false; }
 })();
-/** @type {()=> void} */
-let error;
-const // global consts
-	headline= "bs: Build system based on executables",
-	css= log.css`
-		.error { color: lightred; }
-		.code { color: blue; }
-		.code::before, .code::after { content: "\`"; }
-		.tab::before { content: "	"; }
-		.info {}
-		.headline { color: magenta; }
-		.highlight { color: lightyellow; }
-		.script::before { content: "λ "; }
-		.cwd::before { content: "⌂ "; }
-	`,
-	{ version, name, homepage, description }= pipe( readFileSync, JSON.parse )(__dirname+"/package.json");
-let // bs folder
-	folder_root;
+const { css, fc, info, help }= require("./src/consts.js");
+function logHead(wd_bs){
+	if(stdoutIsFIFO) return;
+	let wd= join(wd_bs, "../");
+	const h= homedir();
+	if(wd.startsWith(h)) wd= join(
+		format("%c~", css.lowpriority),
+		wd.slice(h.length)
+	);
+	return log(
+		"%c%s%c%s",
+		css.cwd,
+		css.lowpriority,
+		wd,
+		join(info.name, "./")
+	);
+}
+const api= require("sade")(info.name)
+	.version(info.version)
+	.describe(help.describe)
 
-const fc= (code, ...rest)=> format("%c"+( !Array.isArray(code) ? code : String.raw(code, ...rest) ), css.code); //format as code
-const catchError= a=> { if(!error) return a; error(a); };
-const api= require("sade")(name)
-	.version(version)
-	.describe([
-		headline,
-		"",
-		"This script allows you to create build scripts using simple executables¹:",
-		"",
-		`1. Create a ${fc(name)} directory in your repository root`,
-		"2. Implement commands just by adding shell scripts or any other executable (using shellbang¹)",
-		"3. Now, you can run build scrip like:",
-		...[ "bs/build", "./bs/build", "bs build", "bs .run build" ]
-			/* fromat as code */.map(c=> format("%c"+c, css.code+css.tab))
-			/* join with "or" */.reverse().map((s, i)=> i ? s+" or" : s).reverse(),
-		"",
-		"So, this script is not neccessary, but it provides some helpers:",
-		`1. You can call executables without extensions (for example ${fc`${name}/test.py`} ⇔ ${fc`${name} test`})`,
-		`2. You can use completion, see ${fc`.completion`} command`,
-		`3. This utility can find current or any parent folder containing ${fc(name)} directory`,
-		"",
-		"To point out:",
-		`1. To prevent colision all ${fc(name)} commands starts with ${fc`.`} (e.g. ${fc`.ls`})`,
-		`2. It is a good practice to distinc non-commands from commands (eg. with preposition ${fc`.`}, ${fc`_`}, …)`,
-		"",
-		"Known pitfalls:",
-		`1. The ${fc(name)} tries to find root folder and uses it as cwd, so ${fc`${name} command ./file`} can works unexpectedly!`,
-		"",
-		"Notes:",
-		`[1] use ${fc`chmod +x`} and shebang² like ${fc`#!/usr/bin/env node`} (similarly for bash, …)`,
-		"[2] https://en.wikipedia.org/wiki/Shebang_(Unix)"
-	].map(linesToMaxLength(65)))
-
-	.command(
-		".run [script]",
-		"Run the given build executable "+format("%c (default when `bs [script]`)", css.highlight),
-		{ default: true }
-	)
+	.command(".run [script]", help.run, { default: true })
 	.action(run)
 
-	.command(
-		".ls [filter]",
-		"Lists all available executables"+format("%c (default when only `bs`)", css.highlight),
-	)
+	.command(".ls [filter]", help.ls)
 	.action((filter)=> {
+		const folder_root= folderRoot();
 		const list= ls({ is_out: true });
 		if(!list.length){
 			if(!folder_root){
-				log("%cNo `bs` for current directory: %c%s", css.error, css.unset, pwd());
-				log("You may want to run %cbs .mkdir%c", css.code, css.unset);
+				log("%cNo `bs` for current directory: %c%s.", css.error, css.unset, pwd());
+				log("You may want to run %cbs .mkdir%c.", css.code);
 			} else
-				log(`No executables found in '${folder_root}'.`);
-			log(`Run %c${name} --help%c for more info.`, css.code, css.unset);
-			return process.exit(1);
+				log("No executables found in '%s'.", folder_root);
+			log("Run %c%s --help%c for more info.", css.code, css.unset, info.name);
+			return exit(1);
 		}
-		(filter ? list.filter(x=> x.script.includes(filter)) : list)
-			.forEach(lsPrintNth);
-		log("\nFor more info use %cbs .cat%c", css.code, css.unset);
-		return process.exit(0);
+		const list_final= filter ? list.filter(x=> x.script.includes(filter)) : list;
+		list_final.forEach(lsPrintNth);
+		if(stdoutIsFIFO) return exit(0);
+
+		log();
+		if(!list_final.length){
+			log("%cNo executables found.", css.error);
+		} else {
+			if(list_final.length>10 && !filter)
+				log("You can filter executables using %cbs .ls filter%c.", css.code);
+			log("You may find more info using %cbs .grep [name|filter]%c.", css.code);
+		}
+		return exit(0);
 	})
 
-	.command(".mkdir [root]", [ "This initializes the projects bs directory",
-		`With ${fc`root`} folder defaults to ${fc`.`}.` ])
+	.command(".mkdir [root]", help.mkdir)
 	.action(init)
 
-	.command(".readme", "This is primarly used for update current bs/README.md content.")
+	.command(".readme", help.readme)
 	.action(init)
 
-	.command(".cat [filter]", "This prints bs/README.md content")
+	.command(".grep [filter]", help.cat)
+	.alias(".cat")
 	.action(cat)
 
-	.command(".completion <shell>", [ "Register a completions for the given shell",
-		`This provides completions for ${fc`bs`} itself and available executables`,
-		"and argumnets for executables if specify in corresponding config file.",
-		"",
-		"To allow completions:",
-		`Just add ${fc`eval "$(${name} .completion bash)"`} to your ${fc`.bashrc`}` ])
+	.command(".completion <shell>", help.completion)
 	.action(completion);
 api.parse(passBuildArgs());
 
+/** @returns {{ content: string[], found: { [script: string]: number }, path: string, exists: boolean, h_level: number }} */
 function readReadme(path_bs){
 	const path= join(path_bs, "README.md");
-	const content= existsSync(path) ? readFileSync(path, "utf8").split("\n") : [
-		"# "+headline,
-		`This project uses [${homepage.slice("https://github.com/".length)}: ${description}](${homepage}).`,
+	const exists= existsSync(path);
+	const content= exists ? readFileSync(path, "utf8").split("\n") : [
+		"# "+info.headline,
+		`This project uses [${info.homepage.slice("https://github.com/".length)}: ${info.description}](${info.homepage}).`,
 		"",
 		"## Available executables",
 		"",
 	];
 	let h_level= 3; // defaults to ###
 	const found= content
-		.flatMap((line, i)=> line.match(/^#+ `?(.\/)?bs\/(.*)/) ? [ [ content[i], i ] ] : []) //`
+		.flatMap((line, i)=> line.match(new RegExp("^#+ `?(.\/)?"+info.name+"/(.*)")) ? [ [ content[i], i ] ] : []) //`
 		.map(function([ l, i ]){
 			h_level= l.indexOf(" ");
-			const script_start= l.indexOf("bs/")+3;
+			const script_start= l.indexOf(info.name+"/")+3;
 			// index of space or ` or line end
 			const script_end= /( |`|$)/g.exec(l.slice(script_start));
 			const end_adjust= script_start + ( script_end[0]!=="" ? 0 : 1 );
@@ -134,39 +104,43 @@ function readReadme(path_bs){
 			return [ key, i ];
 		});
 
-	return { path, content, found: Object.fromEntries(found), h_level };
+	return { path, exists, content, found: Object.fromEntries(found), h_level };
 }
 
 function init(root= pwd()){
 	const is_init= argv.slice(2)[0]===".mkdir";
-	const folder_root_local= is_init ? join(root, name) : ( loadBS(), folder_root );
+	const folder_root_local= is_init ? join(root, info.name) : folderRoot();
 	if(!existsSync(folder_root_local)) mkdirSync(folder_root_local);
-	console.log("Folder: "+folder_root_local);
+	log("%cFolder:", css.highlight, folder_root_local);
 	const readme= readReadme(folder_root_local);
 	const execs= listExecutables(folder_root_local, 0).map(e=> e.replace(folder_root_local+"/", ""));
 	const execs_known= Object.keys(readme.found);
-	console.log("Executables: "+execs.join(", "));
+	log("%cExecutables:", css.highlight, execs.join(", "));
 	const execs_add= execs
 		.filter(e=> !execs_known.includes(e))
 		.map(e=> "#".repeat(readme.h_level)+" bs/"+e);
-	writeFileSync(readme.path, readme.content.join("\n"));
-	if(execs_add.length)
-		console.log("Missing in README: \n```markdown\n"+execs_add.join("\n")+"\n```");
-	console.log("Readme: "+readme.path);
-	process.exit(0);
+	if(!readme.exists)
+		writeFileSync(readme.path, readme.content.join("\n"));
+	log(
+		"%cReadme%s:",
+		css.highlight,
+		readme.exists ? "" : " (created)",
+		readme.path
+	);
+	if(execs_add.length){
+		log("%cMissing in README:", css.highlight);
+		log("```markdown\n%c"+execs_add.join("\n")+"\n%c```", css.headline, css.unset);
+	}
+	return exit(0);
 }
 function cat(filter){
-	loadBS();
-	if(!folder_root){
-		log("%cNo `bs` directory found", css.error);
-		return process.exit(1);
-	}
-	log("%c"+folder_root, css.cwd);
+	const folder_root= folderRoot(true);
+	logHead(folder_root);
 	const readme= join(folder_root, "README.md");
 	const readme_content= existsSync(readme) ? readFileSync(readme, "utf8") : "";
 	if(!readme_content){
 		log("%cNo `README.md` (content) found in `bs` directory", css.error);
-		return process.exit(1);
+		return exit(1);
 	}
 	let lines= [], filtered= false;
 	for(const line of readme_content.split("\n")){
@@ -181,17 +155,25 @@ function cat(filter){
 		}
 		if(filtered) lines.push(line);
 	}
+	if(!lines.length){
+		if(!filter) log("%cReadme is empty!", css.error);
+		else log("%cNo docs for given filter/name.", css.error);
+		return exit(1);
+	}
 	lines.forEach(function echoLine(line){
 		if(line.trim().startsWith("#"))
 			return log("%c"+line, css.headline);
 		log(line);
 	});
-	process.exit(0);
+	if(!filter && !stdoutIsFIFO && lines.length>10)
+		log("\nYou can filter docs using %cbs .grep [name|filter]%c.", css.code);
+	return exit(0);
 }
+/** @returns { { script: string, name: string, docs: string }[] } */
 function ls({ is_out= false }= {}){
-	loadBS();
+	const folder_root= folderRoot();
 	if(!folder_root) return [];
-	if(is_out) log("%c"+folder_root, css.cwd);
+	if(is_out) logHead(folder_root);
 	const { content, found }= readReadme(folder_root);
 
 	return listExecutables(folder_root, 0)
@@ -216,68 +198,70 @@ function lsPrintNth({ script, docs }){
 	log("%c"+out, css.script);
 }
 function run(script){
-	loadBS();
+	const folder_root= folderRoot(true);
 	const args= argv.slice(2);
 	
 	if(args[0]===".run") args.shift();
 	if(!args.length) return api.tree[".ls"].handler();
-
 	else args.shift();
-	catchError();
-	// TODO: what about `./` in args when cwd≠folder_root/..
-	process.chdir(folder_root.replace(/\/bs$/, ""));
-	script= "bs/"+script;
+	const wd_current= pwd();
+	const wd_root= join(folder_root, "..");
+	chdir(wd_root);
+	script= info.name+"/"+script;
 	if(!existsSync(script) || !statSync(script).isFile()){
 		const candidate= listExecutables(script.slice(0, script.lastIndexOf("/")), 0)
 			.find(f=> f.startsWith(script) && f[script.length]===".");
 		if(candidate)
 			script= candidate;
 	}
-	if(!stdoutIsFIFO)
-		log("%c%s", css.cwd, folder_root);
+	if(!stdoutIsFIFO) logHead(folder_root)
 	if(!isExecutable(script)){
-		log(`%c'${script}' doesn't exist or is not executable`, css.error);
+		log(`%c'${script}' doesn't exist or is not executable.`, css.error);
 		const found= ls().filter(f=> f.script.includes(script.slice(3)));
 		if(found.length){
 			log("\nYou may want to run:");
 			found.forEach(lsPrintNth);
 		}
-		return process.exit(1);
+		return exit(1);
 	}
 	
 	if(!stdoutIsFIFO)
-		lsPrintNth({ script });
+		lsPrintNth({ script: script.slice(info.name.length+1) });
 	const { spawn }= require("node:child_process");
-	return spawn(script, args, { stdio: "inherit" })
+	return spawn(
+		script,
+		args.map(a=> [ "./", "../" ].find(p=> a.startsWith(p)) ? relative(wd_root, join(wd_current, a)) : a),
+		{ stdio: "inherit" }
+	)
 		.on("exit", function onexit(exit_code, signal){
 			if(typeof exit_code === 'number')
-				return process.exit(exit_code);
+				return exit(exit_code);
 			
 			log("%cUnknown error", css.error);
-			process.kill(process.pid, signal)
+			kill(pid, signal)
 		});
 }
 function completion(shell){
 	const { completionBash, completionRegisterBash }= require("./src/completion.js");
 	if("bash"===shell)
-		return completionRegisterBash(name);
+		return completionRegisterBash(info.name);
 	if("bash--complete"===shell)
 		return completionBash(
 			{ api, completionScript, ls: ()=> ls().map(r=> r.name) },
 			argv.slice(4)
 		);
 	log("Unknown shell: "+shell);
-	process.exit(1);
+	exit(1);
 }
 function completionScript(name){
 	let bsrc;
-	if(!folder_root) loadBS();
+	const folder_root= folderConfig();
 	const end_empty= { "subcommands": [] };
 	if(!folder_root) return end_empty;
 
 	const { readdirSync }= require("node:fs");
 	for(const file of readdirSync(folder_root)){
-		if(!file.startsWith(".bsrc")) continue;
+		if(!file.startsWith("completion")) continue;
 		bsrc= join(folder_root, file);
 		break;
 	}
@@ -290,24 +274,4 @@ function completionScript(name){
 	} catch(e){
 		return end_empty;
 	}
-}
-
-function loadBS(){
-	if(!folder_root) folder_root= findBS();
-}
-function findBS(cwd= pwd()){
-	const folder_root= "/bs"; // allow change?
-	let candidate= cwd.replace(/\/$/, "");
-	while(!existsSync(candidate+folder_root)){
-		const last_slash= candidate.lastIndexOf("/");
-		if(last_slash < 0){
-			error= ()=> {
-				log("%cNo `bs` for current directory: %c%s", css.error, css.unset, cwd);
-				return process.exit(1);
-			};
-			return null;
-		}
-		candidate= candidate.slice(0, last_slash);
-	}
-	return candidate+folder_root;
 }
